refactor: move tests setup
